Make FIXED PMC trigger general
This commit enables the selection of a fixed PMC as the trigger for the
sampling period. Before it was hardcoded to FIXED 1. All the code has
been adapted to be dynamically configured.
The PMI handler has been provided with a mechanism to avoid system
hang. It tries to adjust the reset_period every time there are too many
PMI occurrences between two context switches.
The adjustment of the reset_period should be synchronized among all
the cores, avoiding subsequent modifications. The ack of the APIC EOI
is now part of the PMI handler, and it is extended to all handlers
running upon a fast irq entry.