Fix system hang by PMI
Generally, the system reacts well when placing some printk inside
an IRQ handler, but in particular cases, this leads to a system hang.
In our experimental setup, this happens when the system generates
a lot of threads (fork) - osbench: thread generations test.
To overcome this problem, all the prints inside the PMI handler have
been removed.

This commit implies minor fixes.