Introduce Intel PT support
This commit adapts the Recode module and related tools to interact
with the Intel PT support. This is a basic implementation to test the
PTWRITE assembly instruction on supporting CPUs (Goldmond plus).
Almost all of the logic is hardcoded and can just activate and
deactivate the PT feature to log packets into a 4K buffer.