Refactor and cleanup
All PMU's related stuff has been moved into a dedicated directory.
The statistics utilities now provide a more straightforward interface.

It is possible to compile the module with debug support so that the
pr_debug calls show their contents (make debug).

There is still a bug when the frequency is high. A spurious NMI fires
and the handler can not understand which is the source. Even though
the PMI handler jumps to the end without performing any action,
the PMI generation stops after that event. So we can infer that
the NMI is generated by some PMC-related condition that we should check.