Introduce multiplexing support
This commit is breaking. To add the PMC multiplexing strategy and the
preliminary support for the events sets switching, the core of the event
management has been redesigned.

The struct hw_events (per-CPU) holds all the events for that set.
At the moment, it is hardcoded to be one, but in the next release,
it will be a list.

The multiplexing strategy is quite essential. Each PMI divides the
quantum for the required events and sets up the PMCs but doesn't adjust
the collected value. Thus the result is not normalized for the general
sampling period, and the PMC with PMI enabled.

The logging system has changed. It has a raw memory buffer that is used
to allocate memory for the samples to address the constraint related to
the variable size of the sample (related to the events set of the sample).
The buffer is safe to be used with one writer and one reader
(even simultaneously).

The PMI callback is called only when the sample is ready.
Hence there may be more PMIs than generated samples.

Samples can be read by the proc interface as before, but accessing
/proc/recode/cpus/cpu* does not give the sample header. Instead, only
a mask (as the first element of the row) has to be matched with the
metadata available under /proc/recode/hw_events. In this way,
it is possible to deal with heterogeneous samples.

Most of the PMU-related data is packed inside a per-CPU variable
which allows handling each core separately.