Scale multiplexing values and set multiple hw_events set.
Each core has its own reset_period, which is computed by the number of
hw_events currently used and the available PMCs. So, if the multiplexing
is required, the overall reset_period stays equal on all cores, but the PMIs
reset period to generate partial samples may change.

The hw_events are placed inside a global array to allow the coexistence
of different sets on different cores. This structure paves the way for the
TMA strategy.

Clean up some commented code lines.