Add modular logic and refactor module structure.
This massive commit extends the modular structure of the module by
adding unique logic that can be enabled at compile-time and provide
a set of particular capabilities to ReCode.

To allow such a modular infrastructure, ReCode provides a dedicated
'callback' structure to set the per-logic callbacks (when required),
then extending the baseline behavior.

At the moment, only TMA and SECURITY logics are defined.

SECURITY logic is the plugin version of the previous version of ReCode.
Furthermore, the proc interface is adapted.

Proc functionalities are adapted to work both with older (< 5.6.0)
and newer kernel versions.

The multiplexing is improved by identifying context where it is not
required. In that case, a per-CPU variable is used to save the last
PMCs values, and the PMC setup is not performed.

The ReCode python script is adapted to the new functionalities.
Bugfix of wrong path selection while running multiple commands.